Add support for new Go error functions
Since the implementation found in https://godoc.org/golang.org/x/xerrors
will be present in go1.13, it is better to make this library compatible
from day one. However, in order to be compatible with older Go versions,
as promised in https://github.com/gbrlsnchs/jwt/issues/26, this
compatibility layer is required.